falta emprolijar, revisar todo porlas

+++ TP1/src/calculos.c
@@ -1,14 +1,14 @@
-/*
- * calculos.c
- *
- *  Created on: 8 sep. 2020
- *      Author: micavazzana
- */
-
 #include <stdio.h>
 #include <stdlib.h>
 #include "calculos.h"
 
+/**
+ * \brief Realiza la operacion matematica SUMA
+ * \param float operandoUno, primer operando
+ * \param float operandoDos, segundo operando
+ * \param float *pResultado, puntero a donde de guardara el resultado de la operacion
+ * \return (-1) ERROR (0) EXITO
+ */
 int sumar(float operandoUno, float operandoDos, float *pResultado) {
 
 	int retorno = -1;
@@ -21,6 +21,13 @@ int sumar(float operandoUno, float operandoDos, float *pResultado) {
 	return retorno;
 }
 
+/**
+ * \brief Realiza la operacion matematica RESTA
+ * \param float operandoUno, primer operando
+ * \param float operandoDos, segundo operando
+ * \param float *pResultado, puntero a donde de guardara el resultado de la operacion
+ * \return (-1) ERROR (0) EXITO
+ */
 int restar(float operandoUno, float operandoDos, float *pResultado) {
 
 	int retorno = -1;
@@ -33,6 +40,13 @@ int restar(float operandoUno, float operandoDos, float *pResultado) {
 	return retorno;
 }
 
+/**
+ * \brief Realiza la operacion matematica DIVISION
+ * \param float operandoUno, primer operando
+ * \param float operandoDos, segundo operando
+ * \param float *pResultado, puntero a donde de guardara el resultado de la operacion
+ * \return (-1) ERROR (0) EXITO
+ */
 int dividir(float operandoUno, float operandoDos, float *pResultado) {
 	int retorno = -1;
 
@@ -44,6 +58,13 @@ int dividir(float operandoUno, float operandoDos, float *pResultado) {
 	return retorno;
 }
 
+/**
+ * \brief Realiza la operacion matematica MULTIPLICACION
+ * \param float operandoUno, primer operando
+ * \param float operandoDos, segundo operando
+ * \param float *pResultado, puntero a donde de guardara el resultado de la operacion
+ * \return (-1) ERROR (0) EXITO
+ */
 int multiplicar(float operandoUno, float operandoDos, float *pResultado) {
 	int retorno = -1;
 
@@ -55,7 +76,13 @@ int multiplicar(float operandoUno, float operandoDos, float *pResultado) {
 	return retorno;
 }
 
-int calcularFactorial(float operando, long long int *pResultado) {
+/**
+ * \brief Calcula el FACTORIAL de un numero
+ * \param float operando, operando ingresado
+ * \param long int *pResultado, puntero a donde de guardara el resultado de la operacion
+ * \return (-1) ERROR (0) EXITO
+ */
+int calcularFactorial(float operando, long int*pResultado) {
 
 	int retorno = -1;
 	long long int auxFactorial=1;
