+++ TP1/src/calculos.h
@@ -8,39 +8,14 @@
 #ifndef CALCULOS_H_
 #define CALCULOS_H_
 
-/**
- * \brief
- * \param
- * \return
- */
 int sumar(float operandoUno, float operandoDos, float *pResultado);
 
-/**
- * \brief
- * \param
- * \return
- */
 int restar(float operandoUno, float operandoDos, float *pResultado);
 
-/**
- * \brief
- * \param
- * \return
- */
 int dividir(float operandoUno, float operandoDos, float *pResultado);
 
-/**
- * \brief
- * \param
- * \return
- */
 int multiplicar(float operandoUno, float operandoDos, float *pResultado);
 
-/**
- * \brief
- * \param
- * \return
- */
-int calcularFactorial(float operando, long long int *pResultado);
+int calcularFactorial(float operando, long int *pResultado);
 
 #endif /* CALCULOS_H_ */
